use a central playwright config
